+++ RELEASES.md
@@ -29,6 +29,7 @@ Upcoming
 - Show "Test a PipeLine" pipeline logging (bugfix)
 - Hide strings in IBIS console messages too (same as in logging)
 - Move config warning "Element not in list of available root elements" to startup
+- Larva: add possibility to overwrite windiff command